refactor(workspace): extract types to workspace-contract
