+++ src/tool-session.ts
@@ -1,64 +1,5 @@
-import type { ResolvedFeatureFlags } from "./feature-flags-contract";
 import { MAX_CONSECUTIVE_TOOL_FAILURES, MAX_TOTAL_STEPS, MAX_TURN_STEPS, TOOL_TIMEOUT_MS } from "./lifecycle-constants";
-import type { ActiveSkill } from "./skill-contract";
-import type { ToolCache } from "./tool-contract";
-import type { WorkspaceProfile } from "./workspace-profile";
-
-export type ToolCallStatus = "succeeded" | "failed";
-
-export type ToolCallRecord = {
-  toolName: string;
-  args: Record<string, unknown>;
-  taskId?: string;
-  resultHash?: string;
-  exitCode?: number;
-  status: ToolCallStatus;
-};
-
-export type SessionFlags = {
-  turnStepCount?: number;
-  turnStepLimit?: number;
-  totalStepLimit?: number;
-};
-
-export type ToolErrorSummary = { message: string; code?: string; kind?: string };
-
-export type PreToolContext = { toolId: string; toolCallId: string; args: Record<string, unknown> };
-export type PostToolContext =
-  | {
-      toolId: string;
-      toolCallId: string;
-      args: Record<string, unknown>;
-      status: "succeeded";
-      result: unknown;
-    }
-  | {
-      toolId: string;
-      toolCallId: string;
-      args: Record<string, unknown>;
-      status: "failed";
-      error: ToolErrorSummary;
-    };
-export type EffectOutput = { append?: string };
-
-export type SessionContext = {
-  callLog: ToolCallRecord[];
-  taskId?: string;
-  flags: SessionFlags;
-  writeTools: ReadonlySet<string>;
-  toolTimeoutMs?: number;
-  cache?: ToolCache;
-  featureFlags?: ResolvedFeatureFlags;
-  consecutiveFailures: Map<string, number>;
-  maxConsecutiveToolFailures?: number;
-  onDebug?: (event: `lifecycle.${string}`, data: Record<string, unknown>) => void;
-  onBeforeTool?: (ctx: PreToolContext) => EffectOutput | undefined;
-  onAfterTool?: (ctx: PostToolContext) => EffectOutput | undefined;
-  onBeforeToolAsync?: (ctx: PreToolContext) => Promise<void>;
-  onAfterToolAsync?: (ctx: PostToolContext) => Promise<void>;
-  workspaceProfile?: WorkspaceProfile;
-  activeSkills?: ActiveSkill[];
-};
+import type { SessionContext, SessionFlags, ToolCallRecord, ToolCallStatus } from "./tool-contract";
 
 export function createSessionContext(taskId?: string, writeTools: ReadonlySet<string> = new Set()): SessionContext {
   return {
